fix: all clippy lints clean; parallel tileset/effect; rayon errors propagated; validate --json; CONTRIBUTING
Clippy (cargo clippy --all-targets -- -D warnings now passes):
- redundant_closure: |x| f(x) → f throughout anim, casc, sprite, validation
- push_str single char: .push_str(",") → .push(',')
- empty_line_after_doc_comment: removed blank lines between /// and items
- div_ceil: (a+b-1)/b → a.div_ceil(b) in anim, sprite, main
- needless_return: removed explicit return at function ends
- useless_vec / useless_use_of_vec: vec![x] → array/slice where possible
- map_or simplifications, matches! macro, manual range contains
- DWORD → Dword (capitalized acronym)
- loop var used to index → iter_mut().enumerate()
- vacuously-true unsigned >= 0 assertions removed
- else { if } → else if collapsing in validate run
- #[allow(clippy::too_many_arguments)] on cmd_extract_anim (8 args)

main.rs features:
- Rayon errors: par_iter().for_each() → .map().collect::<Vec<Result>>()
  so PNG conversion failures are now reported after parallel phase
- Parallel tileset/effect: phase-1 sequential extract, phase-2 parallel
  file writes via par_iter (CascArchive stays in phase 1)
- validate run --json: machine-readable JSON output with per-file results

CONTRIBUTING.md: full rewrite with current prerequisites, test commands,
CI steps, project structure, and contribution workflow

+++ src/casc/encryption.rs
@@ -1,8 +1,7 @@
 /// CASC Encryption and Protection Handling
-/// 
+///
 /// This module provides functionality for handling encrypted and protected files
 /// in CASC archives, including decryption key management for legitimate installations.
-
 use std::collections::HashMap;
 use std::path::{Path, PathBuf};
 use thiserror::Error;
@@ -260,7 +259,7 @@ impl EncryptionHandler {
         // Generate a 32-byte key from the hash
         let mut key = Vec::with_capacity(32);
         for i in 0..4 {
-            let bytes = ((hash >> (i * 8)) as u64).to_le_bytes();
+            let bytes = (hash >> (i * 8)).to_le_bytes();
             key.extend_from_slice(&bytes);
         }
